Refactor create-purplet Fixes #27 and #44
