+++ packages/create-purplet/coolio/.gitignore
@@ -1,0 +1,22 @@
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
+# dependencies
+node_modules
+
+# testing
+coverage
+
+# build
+.purplet
+dist/
+
+# misc
+.DS_Store
+*.pem
+
+# debug
+*debug*.log*
+
+# env files
+.env
+*.env.*
