+++ packages/create-purplet/coolio/.prettierignore
@@ -1,0 +1,4 @@
+dist
+.purplet
+LICENSE
+build
