+++ packages/create-purplet/coolio/.prettierrc
@@ -1,0 +1,7 @@
+{
+  "arrowParens": "avoid",
+  "printWidth": 100,
+  "singleQuote": true,
+  "tabWidth": 2,
+  "trailingComma": "es5"
+}
