+++ packages/create-purplet/coolio/purplet.config.js
@@ -1,0 +1,2 @@
+const config = {};
+export default config;
