+++ packages/create-purplet/coolio/src/lib/utils.js
@@ -1,0 +1,8 @@
+/**
+ * @param {string} str
+ *
+ * @returns {string}
+ */
+export function capitalize(str) {
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
