+++ packages/create-purplet/coolio/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./.purplet/tsconfig.json"
+}
