+++ packages/create-purplet/src/options/eslint.ts
@@ -1,0 +1,9 @@
+import { ProjectConfig } from '../types';
+
+export async function handleESLint(config: ProjectConfig) {
+  return config.eslint
+    ? (async () => {
+        //
+      })()
+    : Promise.resolve();
+}
